+++ odoo/local-src/sf_account/models/account_invoice.py
@@ -9,7 +9,7 @@ class AccountInvoice(models.Model):
 
     def _compute_sale_orders(self):
         for invoice in self:
-            self.order_ids = invoice.invoice_line_ids.mapped(
+            invoice.order_ids = invoice.invoice_line_ids.mapped(
                 'sale_line_ids').mapped('order_id')
 
     @api.multi
@@ -39,3 +39,5 @@ class AccountInvoice(models.Model):
         string='Partner Reference',
         help='Invoice number of the partner'
     )
+    linked_partner_bank_id = fields.Many2one(
+        related='partner_id.bank_ids.linked_partner_id', readonly=True)