+++ HISTORY.rst
@@ -31,6 +31,7 @@ latest (unreleased)
 * BSSFL-365: Import repair line
 * BSSFL-402: Add date delivered field on DO
 * BSSFL-272 : Update res.company to include account cutoff settings
+* BSSFL-417: Pay PO to another partner bank account
 
 **Bugfixes**